renames InstantiateContext -> useInstantiation
