+++ src/features/instance/InstanceContext.tsx
@@ -15,7 +15,7 @@ import { Loader } from 'src/core/loading/Loader';
 import { useHasPendingScans } from 'src/features/attachments/useHasPendingScans';
 import { cleanUpInstanceData } from 'src/features/instance/instanceUtils';
 import { useProcessQuery } from 'src/features/instance/useProcessQuery';
-import { useInstantiation } from 'src/features/instantiate/InstantiationContext';
+import { useInstantiation } from 'src/features/instantiate/useInstantiation';
 import { useInstanceOwnerParty } from 'src/features/party/PartiesProvider';
 import { useNavigationParam } from 'src/features/routing/AppRoutingContext';
 import { buildInstanceDataSources } from 'src/utils/instanceDataSources';
